Rename layout package to mazebuilder
